requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==4.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.1
google-auth==2.49.1
google-genai==1.68.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.2
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.48
starlette==0.52.1
tenacity==9.1.4
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.42.0
websockets==16.0